Move unit tests to unit dir
