+++ tests/unit/test_download_file_slice.py
@@ -10,7 +10,7 @@ import os
 
 from urllib3.exceptions import NewConnectionError
 
-import tests.conftest as common
+import tests.unit.conftest as common
 from pyega3.libs.data_file import DataFile
