+++ tests/unit/test_merge_bin_files.py
